app/requirements.txt
flask
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation-flask
deprecated